modularized/requirements.txt
ultralytics==8.3.162
Flask
opencv-python-headless
numpy
Pillow
bleach
python-dotenv
psycopg2-binary